+++ app/config/development/application.php
@@ -1,14 +1,17 @@
 <?php
 
-return array(
-    'database' => array(
+return [
+    'database' => [
         'adapter' => 'Mysql',
         'host' => 'localhost',
         'username' => 'root',
         'password' => '111',
         'dbname' => 'yona-cms',
         'charset' => 'utf8',
-    ),
-    'cache' => 'file', // file, memcache
-    'metadata_cache' => 'memory', // memory, apc
-);
+    ],
+    'memcache' => [
+        'host' => 'localhost',
+        'port' => 11211,
+    ],
+    'cache'    => 'file', // file, memcache
+];