+++ app/config/production/application.php
@@ -1,14 +1,17 @@
 <?php
 
-return array(
-    'database' => array(
-        'adapter' => 'Mysql',
-        'host' => 'localhost',
+return [
+    'database' => [
+        'adapter'  => 'Mysql',
+        'host'     => 'localhost',
         'username' => '',
         'password' => '',
-        'dbname' => '',
-        'charset' => 'utf8',
-    ),
-    'cache' => 'file', // file, memcache
-    'metadata_cache' => 'memory', // memory, apc
-);
+        'dbname'   => '',
+        'charset'  => 'utf8',
+    ],
+    'memcache' => [
+        'host' => 'localhost',
+        'port' => 11211,
+    ],
+    'cache'    => 'file', // file, memcache
+];